+++ allink_core/apps/news/admin.py
@@ -45,7 +45,7 @@ class NewsAdmin(PlaceholderAdminMixin, AllinkBaseAdminSortable):
         fieldsets = (
             (None, {
                 'fields': (
-                    'is_active',
+                    'status',
                     'title',
                     'slug',
                     'created',