+++ allink_core/apps/testimonials/admin.py
@@ -11,13 +11,13 @@ Testimonials = get_model('testimonials', 'Testimonials')
 @admin.register(Testimonials)
 class TestimonialsAdmin(AllinkBaseAdminSortable):
     search_fields = ('first_name', 'last_name',)
-    list_display = ('first_name', 'last_name', 'get_categories', 'is_active', 'created', 'modified')
+    list_display = ('first_name', 'last_name', 'get_categories', 'status', 'created', 'modified')
 
     def get_fieldsets(self, request, obj=None):
         fieldsets = (
             (None, {
                 'fields': (
-                    'is_active',
+                    'status',
                     ('first_name', 'last_name'),
                     'lead',
                     'slug',